Dear Diary,
I am so happy to be doing another coding investigation. Absolutely ecstatic.

I'm joking. The only good thing about this is the GIT REPOSITORY, because now I have my
own therapist again! YAY!

Anyway today I re-installed Git, even though I had already done it. I also had to copy some
code that I had done before into this folder, because some shenanigans that I don't fully
understand happened, so I had to abandon ship and start all over again. Thank God Ctrl+C
and Ctrl+V exists, am I right?

Moving on.
I created a flawed conversion table that does the most basic of time period conversions,
because I can't think straight and coding is not my forte. Basically what I did is make
I think what's called a dictionary and I put your basic units of time in there, like years,
months, weeks, you get the idea. Then I made each of those a dictionary, and what I did is
put the necessary conversion in there, if that makes sense.

Example:
Within our dictionary we have a dictionary for years. And in the dictionary for years, we
have weeks. Weeks is equal to 52, because there are 52 weeks in a year.
And so on for the other time units.

And basically, if you multiply the number of years (lets say 4) with how many weeks are in 1
year (52), you get the number of weeks in 4 years (52 * 4).

We can also do this with input. If you input the unit of time, how much of that unit of time,
and the unit of time you want to change to, you can look up the two units of time in the dictionary
and multipy whatever number is is by the number of that unit of time.

Does that make sense? Proabably not, but that's ok.

I know that there are a bunch of accidents waiting to happen with the crude code I have developed,
but this'll do for now because I genuinely could not care less at the moment.

I really really hope I got smarter since last year so I can make code that isn't brain dead, but
I gave up on hope a long long time ago. Let's see where this nightmare takes us.

This is an unusually long commit, but I'm pretty sure most of mine will be like this, because I WILL
Git as a diary. Mark my words.

Okay that's enough for today.