Dear Diary,
GUESS WHO'S BACK.

ME.

AND I HAVE DONE A TINY, TINY BIT OF PART 1!!!

To be honest, I'm already far ahead of what I was last year. I think. I'm so proud of myself
for understanding this way too complicated investigation.

So I managed to calculate simple interest. Yes, I know that's barely nothing, but baby steps.
Also if you expected me to get any furhter than that, you are grossly overestimating my abilities.

So I made a function that calculated simple interest and returned the result. Then I made a bunch
of inputs to get the information and outputted the interest amount along with the total amount in
the account.

THAT RHYMES! (amount, account.)

It's only calculating yearly interest for now because I don't know if my conversion table is
suitable for use yet, but I'm really really happy that I managed to do this.

Maybe later today I can do yearly compound interest as well.

Life sucks, but maybe it won't be so bad, if I keep this up.