Dear Diary, It's been a while since my last commit. Sorry about that, I merely realised how badly I didn't want to do this assignment.
Anyway, this update is to say that I have made a compound interest calculator! Woohoo!
I made a function that calculates compound interest which you just need to put input
into. The thing is, I don't know whether my time conversion thing will work, so the
compounding periods are all custom at the moment, rather than being yearly, quarterly,
etc. So the next thing I'll do is that :')

I just realised that no one is going to keep money in a bank for three months if
the interest only compounds a year, right? So technically, I don't have to worry about
converting from weeks to quarter, or some thing like that.

...right?

I don't have the intellect to come up with something for that, so I'm just going to hope
that what I just said is right.

Now that I think of it, I should probably be asking a teacher instead of telling my
"diary" this.

Eh. Whatever.

But yeah, I managed to do compound interest as well! Who knows, maybe I will be able to
finish this assignment on time (ahhahha. touch wood).

Bye!