Dear Diary, I finished Part 1 of the investigation. I had actually doen that yesterday in class, but I didn't have enough time to commit then. I only just opened my laptop (for purposes at least), hence the late commit. Sorry about that.
So what I did was...

Let me think this was a whole DAY ago.

Ok so what I was having trouble with was the compounding frequency nonsense, and the
conversion of time units. I made an input that asks for the compounding frequency, then
converts it into a number that fits however many time it compounds per the interest rate
time unit. Later(today), I also added an if/else statement so if you decide to have a
custom compounding period its asks you what it is and uses that instead of a time unit.
If that makes sense.

I also had to do the rest of the time conversion stuff so now there are calculations that
assure that everything is the same time unit, which is basically just putting input through
my converter and making it all the same unit.

I'm pretty sure that's it.

I think I'll immediately start working on the next because I have a hunch that I won't finish
this thing on time otherwise. T^T