Dear Diary, Today I did Part 4 of the investigation to the best of my capability. Hopefully what I did was right because otherwise I might kill myself.
I'm not joking.

Any first of all I made a function to, you know, model regular deposits. I put a couple
empty lists in there so then we could output the projections later.

Anyway I used the formula for compound interest with regular deposits within a for loop,
appending values to their respective lists as the loop went on. At the end I zipped the
lists and outputted the results in a table of sorts.

This only works if a length of time is the projection amount, I'll have to think about
how it would work if it was a dollar amount. I relaly hope it's not that different because,
like I said, I'm an inch from jumping off my balcony.

These investigations put WAY too much stress on students. Just saying.

Well anyway that's it for now! Maybe I'll commit again later tonight
(but don't count on it.)