Dear Diary, I was fiddling around with the Git settings and for a second I thought I forgotten to save my work and my heart dropped I was so scared.
We since the last time I committed, a couple of changes have been made.
And I'm sorry, I knew I should've done this before but the whole Git
fiasco and I had more important things on my mind.

Well, anyway.

As of now, I have finished the coding investigation!! YAYYY!

(Except for part 5 but we don't talk about that.)

First thing I did was the projections for part 2. I changed the way my
code works slightly for this, by making the time it takes for the account
to reach its target a variable, and running a for loop with that variable.
The for loop calculates the compound interest for that period of time and
appeands it to a list, which is outputted as the projections.

I also made a function that compares the compound interest for two accounts
for module 3, and I am SO glad you can use functions inside functions.
That was a lifesaver.

Anyway so basically the function has two empty lists, one for each CI
account. A for loop starts and for every time interest is charged, the
compound interest for the account is calculated.And each time is runs,
results are appended to the respective lists.

Then the projections are outputted.

I also changed my deposit function a bit, so that is uses the following
variables: principle, interest, compounding period, time, target amount,
and deposit.

I made it so if t = 0, the time it takes for an account to reach a target
is calculated and then that is used in the for loop.

I also made chnages to the input for part 5, so when it asks for the
target amount, if you give a non-zero number it automatically sets t to
zero, so the function can still work.

I'm going to finish my powerpoint, but if I get the time, I might do part
of the investigation as well.

That's it for today.