Dear Diary, Today, I did Part 5 and therefore completed the investigation!!! YAHOOOOO!!!! You cannot imagine the level of happiness I feel right now.
I basically just used my function for Part 5 and printed out the output for
quarterly, monthly, weekly and daily. When we got to hourly and ten minutely,
the output was too large to present.

But yeah. That's it.

Now I just have to do the powerpoint, and then I can submit! WOOHOO!

ok that is it