This is my final commit.
I would say it has been fun doing this investigation, but that would be
a bare-faced lie.

Even with all the distress this caused me, completing it feels kind of
bittersweet. Maybe it's because I got emotionally attached to my Git
commits.

I feel a thank is in order, Git, for being my personal therapist while I
worked on this assignment. While I do not look forward to the next one,
I sure as hell look forward to using you again.

Goodbye.